% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{clean_columns}
\alias{clean_columns}
\title{clean_columns}
\usage{
clean_columns(df)
}
\arguments{
\item{df}{Uncleaned dataframe as input}
}
\value{
Cleaned Dataframe
}
\description{
Removes prefixes from column names
}
