% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datafinder.R
\name{get_datafinder}
\alias{get_datafinder}
\title{get_datafinder}
\usage{
get_datafinder(jurisdiction, document_type = NULL)
}
\arguments{
\item{jurisdiction}{ID for the jurisdiction}

\item{document_type}{ID for type of document, default value is NULL}
}
\value{
Returns dataframe with the series and years available,
         along with the endpoints to access the data
}
\description{
Get API info for a specific jurisdiction and documentType
}
\examples{
\dontrun{get_datafinder(jurisdiction = 1)}
}
