% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_documentation.R
\name{get_documentation}
\alias{get_documentation}
\title{get_documentation}
\usage{
get_documentation()
}
\value{
Returns a dataframe with information on the source of the
         documentation, the source name, source citation,
         source url and documentation.
}
\description{
Get documentation for projects, including citations.
}
\examples{
\dontrun{get_documentation()}
}
