% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_documents.R
\name{get_documents}
\alias{get_documents}
\title{get_documents}
\usage{
get_documents(
  document_id = NULL,
  jurisdiction_id = NULL,
  date = NULL,
  document_type = 1,
  verbose = 0
)
}
\arguments{
\item{document_id}{ID of the specific document}

\item{jurisdiction_id}{ID for the jurisdiction}

\item{date}{Year(s) of the documents}

\item{document_type}{ID for type of document, default value of 1}

\item{verbose}{prints out the url of the API call, default value of 0}
}
\value{
Returns pandas dataframe with the metadata
}
\description{
Get metadata for documents available in a specific jurisdiction or for a
specific document ID
}
\examples{
\dontrun{get_documents(jurisdiction_id = 38, date = 2022)}
}
