% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint.R
\name{get_endpoint}
\alias{get_endpoint}
\title{get_endpoint}
\usage{
get_endpoint(series, jurisdiction, year, document_type, summary = TRUE)
}
\arguments{
\item{series}{Series ID(s)}

\item{jurisdiction}{ID for the jurisdiction}

\item{year}{Year(s) of data}

\item{document_type}{ID for type of document}

\item{summary}{Return summary instead of document level data
(only one year of data is allowed for document level data),
default value is TRUE}
}
\value{
Returns the endpoint, e.g. '/state-summary' for summary-level state
        data
}
\description{
Get endpoint for a specific series, jurisdiction, year, document_type combo
}
\examples{
\dontrun{get_endpoint(
  series = c(28, 33, 36), jurisdiction = 38,
  year = c(1970, 2003, 2004, 2018), document_type = 1
)}
}
