% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_industries.R
\name{get_industries}
\alias{get_industries}
\title{get_industries}
\usage{
get_industries(
  keyword = NULL,
  label_level = 3,
  label_source = NULL,
  verbose = 0
)
}
\arguments{
\item{keyword}{search for keyword in industry name}

\item{label_level}{NAICS level (2 to 6-digit), default value of 3}

\item{label_source}{classification standard (NAICS, BEA, SOC)}

\item{verbose}{prints out the url of the API call, default value of 0}
}
\value{
Returns pandas dataframe with the metadata
}
\description{
Get metadata for all industries available in a specific jurisdiction
}
\examples{
\dontrun{get_industries()}
}
