% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jurisdictions.R
\name{get_jurisdictions}
\alias{get_jurisdictions}
\title{get_jurisdictions}
\usage{
get_jurisdictions(verbose = 0)
}
\arguments{
\item{verbose}{prints out the url of the API call, default value of 0}
}
\value{
Returns pandas dataframe with the metadata
}
\description{
Get metadata for all or one specific jurisdiction
}
\examples{
\dontrun{get_jurisdictions()}
}
