% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_series}
\alias{get_series}
\title{get_series}
\usage{
get_series(verbose = 0)
}
\arguments{
\item{verbose}{prints out the url of the API call, default value of 0}
}
\value{
Returns pandas dataframe with the metadata
}
\description{
Get series metadata for all or one specific jurisdiction
}
\examples{
\dontrun{get_series()}
}
