% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industries_url.R
\name{industries_url}
\alias{industries_url}
\title{industries_url}
\usage{
industries_url(keyword, label_level, label_source)
}
\arguments{
\item{keyword}{search for keyword in industry name}

\item{label_level}{NAICS level (2 to 6-digit)}

\item{label_source}{classification standard (NAICS, BEA, SOC)}
}
\value{
url as a string
}
\description{
Gets url call for label (formerly industries) endpoint.
}
\examples{
\dontrun{industries_url("test", 323, "NAICS")}
}
