% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurisdictions_url.R
\name{jurisdictions_url}
\alias{jurisdictions_url}
\title{jurisdictions_url}
\usage{
jurisdictions_url()
}
\value{
url as a string
}
\description{
Gets url call for jurisdictions endpoint.
}
\examples{
\dontrun{jurisdictions_url()}
}
