% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_clusters.R
\name{list_clusters}
\alias{list_clusters}
\title{list_clusters}
\usage{
list_clusters(reverse = FALSE)
}
\arguments{
\item{reverse}{reverses the key-value mapping, default value of FALSE}
}
\value{
Returns dictionary containing names of clusters & associated IDs
}
\description{
Fetches the names of the clusters and their associated IDs
}
\examples{
\dontrun{list_clusters()}
}
