% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dates.R
\name{list_dates}
\alias{list_dates}
\title{list_dates}
\usage{
list_dates(jurisdiction_id, document_type = NULL)
}
\arguments{
\item{jurisdiction_id}{ID for the jurisdiction}

\item{document_type}{ID for type of document}
}
\value{
Returns list of dates available for the jurisdiction
}
\description{
Fetches the dates available for the jurisdiction
}
\examples{
\dontrun{list_dates(jurisdiction_id = 38)}
}
