% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_document_types.R
\name{list_document_types}
\alias{list_document_types}
\title{list_document_types}
\usage{
list_document_types(jurisdiction_id = NULL, reverse = FALSE, verbose = 0)
}
\arguments{
\item{jurisdiction_id}{ID for the jurisdiction}

\item{reverse}{reverses the key-value mapping}

\item{verbose}{prints out the url of the API call}
}
\value{
Returns dictionary containing names of documenttypes & associated IDs
}
\description{
Fetches the names of documenttypes and associated IDs
}
\examples{
\dontrun{list_document_types(jurisdiction_id = 65,
         reverse = FALSE, verbose = 0)}
}
