% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_jurisdictions.R
\name{list_jurisdictions}
\alias{list_jurisdictions}
\title{list_jurisdictions}
\usage{
list_jurisdictions(reverse = FALSE)
}
\arguments{
\item{reverse}{reverses the key-value mapping, default value of FALSE}
}
\value{
Returns dictionary containing names of jurisdictions & associated IDs
}
\description{
Fetches names of jurisdictions and associated IDs
}
\examples{
\dontrun{list_jurisdictions(reverse = TRUE)}
}
