% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_series.R
\name{list_series}
\alias{list_series}
\title{list_series}
\usage{
list_series(reverse = FALSE)
}
\arguments{
\item{reverse}{reverses the key-value mapping, default value of FALSE}
}
\value{
Returns dictionary containing names of series and associated IDs
}
\description{
Fetches the names of series and associated IDs
}
\examples{
\dontrun{list_series(reverse = FALSE)}
}
