% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_error.R
\name{print_error}
\alias{print_error}
\title{print_error}
\usage{
print_error(output)
}
\arguments{
\item{output}{key-value mapping of the output}
}
\value{
None
}
\description{
Handle and print out error for invalid API call.
}
\examples{
\dontrun{print_error(list(message = "test"))}
}
