% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_time.R
\name{reading_time}
\alias{reading_time}
\title{reading_time}
\usage{
reading_time(words, workday = 8, workweek = 5, workyear = 50)
}
\arguments{
\item{words}{count of words}

\item{workday}{default value of 8 hour workday}

\item{workweek}{default value of 5 day work week}

\item{workyear}{default value of 50 week work year}
}
\value{
A string detailing how long it takes to read a document based on
how many words the document has. The function assumes an 8 hour work-day,
a 5 day work-week, and a 50 week work-year.

A string detailing how long it takes to read a document based
on how many words the document has. The function assumes an 8 hour work-day,
a 5 day work-week, and a 50 week work-year.
}
\description{
Computes the time it takes to read a document based on the words a document
has.

Computes the time it takes to read a document based on the words a
document has.
}
\examples{
\dontrun{reading_time(1200000, workday = 8,
         workweek = 5, workyear = 50)}
reading_time(1200000, workday=8, workweek=5, workyear=50)
}
